Changing caption parser to create list of bounding boxes.
The bounding boxes now have [0] as the None element instead of 'None'.
Was an experiment in order to remove the error of pickelization of the
dataloader. Turns out since the bounding box lists are of variabl
dimensions, they cannot be packed into a single dataloader. Therefore,
just pass caption ID and fetch it from a loaded json then.

Signed-off-by: Deepan Das <deepandas11@gmail.com>