Added functionality to show localization maps
Can read flickr test data in whatever parse mode and show localization
maps for the entities provided in that caption. The entity phrases are
provided in the flickr dataset.

If parse_mode is default, colocmaps are found for each word. Then,
colocmaps for the words in a given entity phrase is averaged and
visualized.

If parse_mode is phrase, colocmaps for each phrase is evaluated and

Signed-off-by: deepandas11 <deepandas11@gmail.com>